requirements.txt
bsp-tool==0.4.0
PyYAML==6.0
vpk==1.4.0
